policy: move ruleSlice functions to new rules.go
No functional change is intended here.

Signed-off by: Ian Vernon <ian@cilium.io>
